+++ backend/src/WebApi/Features/Expenses/Commands/Create/CreateExpenseCommand.cs
@@ -4,7 +4,7 @@ namespace WebApi.Features.Expenses.Commands.Create;
 
 public record CreateExpenseCommand(
     string Description,
-    double Amount,
+    decimal Amount,
     string Category,
     DateTime Date
 ) : IRequest<Guid>;
