+++ backend/src/WebApi/Features/Expenses/Commands/Update/UpdateExpenseCommand.cs
@@ -5,7 +5,7 @@ namespace WebApi.Features.Expenses.Commands.Update;
 public record UpdateExpenseCommand(
     Guid Id,
     string Description,
-    double Amount,
+    decimal Amount,
     string Category,
     DateTime Date
 ) : IRequest<bool>;
